Struct ltspice_parse::results::VariableResult

pub struct VariableResult<'a> {
    // some fields omitted
}

This struct contains all data for one result of a simulation step. The data is accessible in different formats

Examples

//Gets the avsolute values in decibel
let decibel_data = result.get_data(&DataType::AbsoluteDecibel);
let starting_freq = decibel_data[0];

Methods

impl<'a> VariableResult<'a>

fn new(variable: &'a SimulationVariable, reals: Vec<f64>, imags: Vec<f64>) -> Self

Creates a new instance of a VariableResult

To get a VariableResult from a .raw-File, the use of SteppedSimulation::from_files is recommended

fn get_data(&self, data_type: &DataType) -> Vec<f64>

Returns the data as a vector.

The contents are determined by the data_type

Example

let imaginary_values = result.get_data(&DataType::Imaginary);

fn diff(raw_data: &Vec<f64>) -> Vec<f64>

Calculates a difference vector of your raw data: d(n) = x(n+1)-x(n)

fn find_peaks(&self, peak_type: Option<PeakType>, data_type: &DataType) -> Vec<usize>

Finds peaks in the data

To find maxima, call it with Some(PeakType::Maximum) To find all peaks, call it with None

Example

let peaks = result.find_peaks(Some(PeakType::Maximum),&DataType::AbsoluteDecibel);

fn get_data_point(&self, data_type: &DataType, index: usize) -> Option<f64>

Returns a single data point converted to the given data type

fn len(&self) -> usize

Returns the length of the data set

fn min(&self, data_type: &DataType) -> (usize, f64)

Returns the minimal value in the data set of the given data type

fn max(&self, data_type: &DataType) -> (usize, f64)

Returns the maximal value in the data set of the given data type

fn avg_normalized(&self, data_type: &DataType, frequency: &VariableResult) -> f64

Returns the average of the values in the given data types. This average is normalized so that the logarithmic scaling of the frequency is not present anymore. This method tries to compute the average in a way, that it would seem, that the frequency is scaled linearly.

fn find_value_near_freq(&self, data_type: &DataType, frequency: &VariableResult, search_freq: f64) -> (usize, f64)

Searches the dataset for the first value after the given search frequency. To search for a value near 500 Hz call

let (position_in_dataset, value) = find_value_near_freq(&DataType::AbsoluteDecibel,&frequency_dataset,500.0)

fn next_value_around(&self, data_type: &DataType, starting_point: usize, offset: f64, maximum: bool) -> (Option<usize>, Option<usize>)

Returns the next occurence of a value higher/lower than the value at the starting point plus/minus the offset in both directions

fn calculate_resonance_penalty(&self, frequency: &VariableResult) -> f64

Used for myself. Not really documented

fn calculate_fitness(&self, frequency: &VariableResult, averages: [f64; 5], deviations: [f64; 5]) -> ([f64; 6], f64, [f64; 6])

Used for myself. Not really documented

fn plot(&self, frequency: &VariableResult, fg: &mut Figure, title: &str, color: &str)

Uses GNUplot to plot the Bode plot (just absolute, no phase)