Struct ltspice_parse::SteppedSimulation

pub struct SteppedSimulation {
    // some fields omitted
}

A set of results of one Stepped Simulation read from a File

Examples

let mut path = PathBuf::from("./simulations/");
let model_name = "simulation1";

path.push(model_name);
    
let results = SteppedSimulation::from_files(
    path.with_extension("raw").as_path(),
    path.with_extension("log").as_path());

This can currently only read results, which are generated with the .ac command. If you want to read transient files, this library has to be extended

Methods

impl SteppedSimulation

fn from_files<P: AsRef<Path>>(path_raw: P, path_log: P) -> Self

Reads the Simulation result form the .raw-File. This also needs the .log-File to get informations about the steps.

fn available_parameters(&self) -> Vec<String>

Returns a vector with all .param parameters as strings.

fn available_variables(&self) -> &Vec<SimulationVariable>

Returns a vector of all available variables which resulted of the simulation

fn get_variable_for_name(&self, name: &str) -> Option<&SimulationVariable>

Returns a variable, which has the given name. If no variable for that name can be found, it returns None

fn available_steps(&self) -> &Vec<Step>

Returns a vector of all steps of the simulation

fn get_values_at(&self, step: &Step) -> Vec<VariableResult>

Returns all VariableResults for one step

fn get_values_for_variable_at(&self, step: &Step, var: &SimulationVariable) -> Option<VariableResult>

Returns the VariableResult for one variable at a given step

fn calculate_fitnesses(&self, var: &SimulationVariable) -> Vec<f64>

Calculates the fitnesses of all steps for a given variable. The fitness is determined by the internal fitness function of the VariableResult. Currently it is not possible to define the fitness function by yourself.

fn find_averages_for_fitness(&self, var: &SimulationVariable) -> ([f64; 5], [f64; 5])

fn find_with_resonance_at(&self, var: &SimulationVariable, res_freq: f64) -> Vec<VariableResult>

Finds all VariableResults, that contain a resonance at the given frequency. This algorithm is not very good at the current moment, as it just takes the frequency ticks with a distance of 2 to determine the borders of the allowed frame.