Struct ltspice_parse::VariableResult

pub struct VariableResult<'a> {
    // some fields omitted
}

A Variable Result

Methods

impl<'a> VariableResult<'a>

fn diff(raw_data: &Vec<f64>) -> Vec<f64>

fn find_peaks(&self, peak_type: Option<PeakType>, data_type: &DataType) -> Vec<usize>

Finds peaks in the data If you want to find maxima, you call it with Some(PeakType::Maximum) If you want to find all peaks, you call it with None

fn get_data(&self, data_type: &DataType) -> Vec<f64>

fn get_data_point(&self, data_type: &DataType, index: usize) -> Option<f64>

fn len(&self) -> usize

fn min(&self, data_type: &DataType) -> (usize, f64)

fn max(&self, data_type: &DataType) -> (usize, f64)

fn avg_normalized(&self, data_type: &DataType, frequency: &VariableResult) -> f64

fn find_value_near_freq(&self, data_type: &DataType, frequency: &VariableResult, search_freq: f64) -> (usize, f64)

fn next_value_around(&self, data_type: &DataType, starting_point: usize, offset: f64, maximum: bool) -> (Option<usize>, Option<usize>)

Returns the next occurence of a value higher/lower than the value at the starting point plus/minus the offset in both directions

fn calculate_resonance_penalty(&self, frequency: &VariableResult) -> f64

fn calculate_fitness(&self, frequency: &VariableResult, averages: [f64; 5], deviations: [f64; 5]) -> ([f64; 6], f64, [f64; 6])

fn plot(&self, frequency: &VariableResult, fg: &mut Figure, title: &str, color: &str)